Log state
Currently the project's frontend was built to support the creation of proposals and an initial staking for the proposal to be created, the logic of the DApp would create a contract, voters would attach and contribute to the contract linked to the proposal. After a predefined timeout elapses then the upvotes are evaluated then an outcome is decided. There are 2 APIs a Contributor and a Proposer, 2 participants a Deployer and an Attacher. Log ended.